+++ simulator/src/utils/utils.js
@@ -6,6 +6,10 @@ var _ = require('lodash');
 var ctx;
 Utils.linkPageWidgets = linkPageWidgets;
 
+function linkPageAllWidgets(page) {
+    page.linkedAllWidgets = linkWidgets(getPageAllInteractiveWidgets(page));
+}
+
 function linkPageWidgets(page) {
     page.linkedWidgets = linkWidgets(getPageInteractiveWidgets(page));
 }
@@ -119,6 +123,7 @@ function linkWidgets(widgetList) {
                 keys.forEach(function (key, index) {
                     linkedWidgetList.push(new LinkedWidget(curWidget.subType,curWidget,index,key.x,key.y));
                 })
+                break;
             default:
                 // linkedWidget.type = curWidget.subType;
                 // linkedWidget.target = curWidget;