+++ changelog.md
@@ -124,3 +124,10 @@ add:
     5. limit duration time to 5000
 
 
+2017.5.15
+fix:
+    1. copy page bug
+    2. mulitiColor progress's cursor not display in simulator
+    3. remove Slide and Switch widget's action
+add:
+    1. add change subLayer zIndex fun